Split ErrorBoundary Component
